refactor(chat): extract membership/invitation/ownership event renderers
Stage 4b (2/2). Moves the remaining 14 event render* functions (leave, join,
invitation, application-response, role-changed, ownership, member-removed) into
the createEventRenderers(ctx) factory in messageEventRenderers.jsx, completing
the event-renderer extraction (29 renderers total). ctx now also carries
Mention, renderAvatar/renderSenderName, getReadByTooltip, currentUserId,
conversationType, teamMembers and searchQuery. Removes 10 now-unused lucide
imports from MessageDisplay. MessageDisplay 3365 → 2489 lines.
