+++ src/components/chat/MessageDisplay.jsx
@@ -9,17 +9,7 @@ import {
   isSyntheticTeam,
 } from "../../utils/userHelpers";
 import {
-  UserPlus,
-  UserMinus,
-  LogOut,
-  PartyPopper,
-  Crown,
-  Shield,
   User,
-  UserCheck,
-  UserSearch,
-  CircleX,
-  FileText,
   Download,
   AlertTriangle,
   Clock,
@@ -1165,929 +1155,37 @@ const MessageDisplay = ({
     renderRoleClosedMessage,
     renderRoleUpdatedMessage,
     renderRoleDeletedMessage,
+    renderLeaveMessage,
+    renderUserLeftLomirMessage,
+    renderMemberRemovedPublicMessage,
+    renderJoinMessage,
+    renderInvitationCancelledMessage,
+    renderInvitationDeclinedMessage,
+    renderApplicationResponseMessage,
+    renderApplicationDeclinedMessage,
+    renderInvitationResponseMessage,
+    renderApplicationCancelledMessage,
+    renderRoleChangedMessage,
+    renderOwnershipTeamMessage,
+    renderOwnershipTransferredMessage,
+    renderMemberRemovedMessage,
   } = createEventRenderers({
+    Mention,
     MentionById,
     TeamMentionById,
     RoleMentionById,
     userMentionOrYou,
     possessiveUserMentionOrYour,
     isCurrentViewer,
+    renderAvatar,
+    renderSenderName,
     highlightEventContent,
+    getReadByTooltip,
+    currentUserId,
+    conversationType,
+    teamMembers,
+    searchQuery,
   });
-
-  // =============================================================================
-  // renderLeaveMessage - Neutral grey theme (pill shape)
-  // =============================================================================
-  const renderLeaveMessage = (message, parsedMessage, isCurrentUser) => {
-    const leaveText = isCurrentUser ? (
-      "You have left the team."
-    ) : (
-      <>
-        <MentionById
-          userId={parsedMessage.userId}
-          name={parsedMessage.userName}
-        />{" "}
-        has left the team.
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <UserMinus size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(leaveText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderUserLeftLomirMessage - Neutral grey theme
-  // =============================================================================
-  const renderUserLeftLomirMessage = (message) => {
-    const leaveText = <>Former Lomir Member has left Lomir.</>;
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <LogOut size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(leaveText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  const renderMemberRemovedPublicMessage = (
-    message,
-    parsedMessage,
-    isCurrentUser,
-  ) => {
-    const isRemovedMemberCurrentUser = isCurrentViewer(parsedMessage.userId);
-    const text = isRemovedMemberCurrentUser ? (
-      "You were removed from the team."
-    ) : isCurrentUser ? (
-      <>
-        You removed{" "}
-        <MentionById
-          userId={parsedMessage.userId}
-          name={parsedMessage.userName}
-        />{" "}
-        from the team.
-      </>
-    ) : (
-      <>
-        <MentionById
-          userId={parsedMessage.userId}
-          name={parsedMessage.userName}
-        />{" "}
-        has been removed from the team.
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <UserMinus size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(text)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  /**
-   * Render a team join message with special formatting
-   * Shows announcement banner + personal message in bubble
-   */
-  const renderJoinMessage = (
-    message,
-    parsedMessage,
-    senderInfo,
-    isCurrentUser,
-    senderId,
-  ) => {
-    const pronoun = isCurrentUser ? "you" : "them";
-    const roleMention = parsedMessage.roleName ? (
-      <>
-        <UserCheck size={16} className="event-inline-icon mx-1" />
-        <Mention name={parsedMessage.roleName} />
-      </>
-    ) : null;
-    const welcomeText = isCurrentUser ? (
-      roleMention ? (
-        <>You joined the team as{" "}{roleMention}. Welcome aboard!</>
-      ) : (
-        <>You joined the team. Welcome aboard!</>
-      )
-    ) : roleMention ? (
-      <>
-        <Mention name={parsedMessage.userName} /> joined the team as{" "}
-        {roleMention}. Say hello to {pronoun}!
-      </>
-    ) : (
-      <>
-        <Mention name={parsedMessage.userName} /> has followed your invite and
-        joined your team. Say hello to {pronoun}!
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--success mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <UserPlus size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(welcomeText)}
-            <PartyPopper size={16} className="event-inline-icon ml-1" />
-          </span>
-        </div>
-
-        {parsedMessage.personalMessage && (
-          <div
-            className={`flex ${
-              isCurrentUser ? "justify-end" : "justify-start"
-            } w-full`}
-          >
-            {!isCurrentUser && renderAvatar(senderInfo, true, senderId)}
-
-            <div className="flex flex-col max-w-[70%]">
-            {!isCurrentUser && (
-                renderSenderName(
-                  senderInfo,
-                  senderId,
-                  "text-xs font-medium mb-1 ml-3",
-                )
-              )}
-
-              <div
-                className={`
-                  rounded-lg p-3 
-                  ${
-                    isCurrentUser
-                      ? "bg-green-100 text-base-content rounded-br-none ml-auto"
-                      : "bg-base-200 rounded-bl-none"
-                  }
-                `}
-              >
-                <p>{renderHighlightedSearchText(parsedMessage.personalMessage, searchQuery)}</p>
-                <div
-                  className={`
-                    flex justify-end items-center text-xs mt-1 
-                    ${
-                      isCurrentUser
-                        ? "text-base-content/60"
-                        : "text-base-content/50"
-                    }
-                  `}
-                >
-                  <span>{formatLocalTime(message.createdAt)}</span>
-                  {
-                    <ReadReceipt
-                      message={message}
-                      isCurrentUser={isCurrentUser}
-                      conversationType={conversationType}
-                      teamMembers={teamMembers}
-                      currentUserId={currentUserId}
-                      getReadByTooltip={getReadByTooltip}
-                    />
-                  }
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-
-        {!parsedMessage.personalMessage && (
-          <div className="text-xs text-base-content/50">
-            {formatLocalTime(message.createdAt)}
-          </div>
-        )}
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderInvitationCancelledMessage - Neutral grey theme
-  // =============================================================================
-  const renderInvitationCancelledMessage = (
-    message,
-    parsedMessage,
-    isCurrentUser,
-  ) => {
-    const messageText = isCurrentUser ? (
-      <>
-        You cancelled your invitation for{" "}
-        {parsedMessage.inviteeId ? (
-          <MentionById
-            userId={parsedMessage.inviteeId}
-            name={parsedMessage.inviteeName}
-          />
-        ) : (
-          <Mention name={parsedMessage.inviteeName} />
-        )}{" "}
-        to join{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />
-        . Want to tell them why in this chat?
-      </>
-    ) : (
-      <>
-        {parsedMessage.cancellerId ? (
-          <MentionById
-            userId={parsedMessage.cancellerId}
-            name={parsedMessage.cancellerName}
-          />
-        ) : (
-          <Mention name={parsedMessage.cancellerName} />
-        )}{" "}
-        cancelled your invitation to join{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />
-        {". "}Want to reach out to them in this chat?
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <CircleX size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderInvitationDeclinedMessage - Neutral grey theme
-  // =============================================================================
-  const renderInvitationDeclinedMessage = (
-    message,
-    parsedMessage,
-    isCurrentUser,
-  ) => {
-    const messageText = isCurrentUser ? (
-      parsedMessage.hasPersonalMessage ? (
-        <>
-          You declined{" "}
-          <MentionById
-            userId={parsedMessage.inviterId}
-            name={parsedMessage.inviterName}
-          />
-          {"'s"} invitation for{" "}
-          <TeamMentionById
-            teamId={parsedMessage.teamId}
-            name={parsedMessage.teamName}
-          />{" "}
-          and added this message:
-        </>
-      ) : (
-        <>
-          You declined{" "}
-          <MentionById
-            userId={parsedMessage.inviterId}
-            name={parsedMessage.inviterName}
-          />
-          {"'s"} invitation for{" "}
-          <TeamMentionById
-            teamId={parsedMessage.teamId}
-            name={parsedMessage.teamName}
-          />
-          . Consider adding a personal message to explain your decision.
-        </>
-      )
-    ) : parsedMessage.hasPersonalMessage ? (
-      <>
-        Your invitation for{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        was declined by{" "}
-        <MentionById
-          userId={parsedMessage.inviteeId}
-          name={parsedMessage.inviteeName}
-        />
-        , who added this message:
-      </>
-    ) : (
-      <>
-        Your invitation for{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        was declined by{" "}
-        <MentionById
-          userId={parsedMessage.inviteeId}
-          name={parsedMessage.inviteeName}
-        />
-        . Want to reach out to them in this chat?
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <CircleX size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderApplicationResponseMessage - Neutral grey theme
-  // =============================================================================
-  const renderApplicationResponseMessage = (
-    message,
-    parsedMessage,
-    senderInfo,
-    isCurrentUser,
-    senderId,
-  ) => {
-    const bannerContent = isCurrentUser ? (
-      <>
-        Your decline response to <Mention name={parsedMessage.applicantName} />
-        {"'s"} application for{" "}
-        <span className="font-medium">{renderHighlightedSearchText(parsedMessage.teamName, searchQuery)}</span>
-      </>
-    ) : (
-      <>
-        Response to your application for{" "}
-        <span className="font-medium">{renderHighlightedSearchText(parsedMessage.teamName, searchQuery)}</span>
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <FileText size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(bannerContent)}
-          </span>
-        </div>
-
-        {parsedMessage.personalMessage && (
-          <div
-            className={`flex ${
-              isCurrentUser ? "justify-end" : "justify-start"
-            } w-full`}
-          >
-            {!isCurrentUser && renderAvatar(senderInfo, true, senderId)}
-
-            <div className="flex flex-col max-w-[70%]">
-              <div
-                className={`
-                  rounded-lg p-3
-                  ${
-                    isCurrentUser
-                      ? "bg-green-100 text-base-content rounded-br-none ml-auto"
-                      : "bg-base-200 rounded-bl-none"
-                  }
-                `}
-              >
-                <p>{renderHighlightedSearchText(parsedMessage.personalMessage, searchQuery)}</p>
-                <div
-                  className={`
-                    flex justify-end items-center text-xs mt-1
-                    ${
-                      isCurrentUser
-                        ? "text-base-content/60"
-                        : "text-base-content/50"
-                    }
-                  `}
-                >
-                  <span>{formatLocalTime(message.createdAt)}</span>
-                  {
-                    <ReadReceipt
-                      message={message}
-                      isCurrentUser={isCurrentUser}
-                      conversationType={conversationType}
-                      teamMembers={teamMembers}
-                      currentUserId={currentUserId}
-                      getReadByTooltip={getReadByTooltip}
-                    />
-                  }
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderApplicationDeclinedMessage - Neutral grey theme
-  // =============================================================================
-  const renderApplicationDeclinedMessage = (
-    message,
-    parsedMessage,
-    isCurrentUser,
-  ) => {
-    const messageText = isCurrentUser ? (
-      parsedMessage.hasPersonalMessage ? (
-        <>
-          You declined{" "}
-          <MentionById
-            userId={parsedMessage.applicantId}
-            name={parsedMessage.applicantName}
-          />
-          {"'s"} application for{" "}
-          <TeamMentionById
-            teamId={parsedMessage.teamId}
-            name={parsedMessage.teamName}
-          />{" "}
-          and added this message:
-        </>
-      ) : (
-        <>
-          You declined{" "}
-          <MentionById
-            userId={parsedMessage.applicantId}
-            name={parsedMessage.applicantName}
-          />
-          {"'s"} application for{" "}
-          <TeamMentionById
-            teamId={parsedMessage.teamId}
-            name={parsedMessage.teamName}
-          />
-          . Consider adding a personal message to explain your decision.
-        </>
-      )
-    ) : parsedMessage.hasPersonalMessage ? (
-      <>
-        Your application to{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        was declined by{" "}
-        <MentionById
-          userId={parsedMessage.approverId}
-          name={parsedMessage.approverName}
-        />
-        {", "}who added this message:
-      </>
-    ) : (
-      <>
-        Your application to{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        was declined by{" "}
-        <MentionById
-          userId={parsedMessage.approverId}
-          name={parsedMessage.approverName}
-        />
-        {". "}Want to reach out to them in this chat?
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <CircleX size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderInvitationResponseMessage - Info blue theme
-  // =============================================================================
-  const renderInvitationResponseMessage = (
-    message,
-    parsedMessage,
-    senderInfo,
-    isCurrentUser,
-    senderId,
-  ) => {
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <FileText size={16} className="event-inline-icon mr-1" />
-            Response to invitation for{" "}
-            <span className="font-medium">{renderHighlightedSearchText(parsedMessage.teamName, searchQuery)}</span>
-          </span>
-        </div>
-
-        {parsedMessage.personalMessage && (
-          <div
-            className={`flex ${
-              isCurrentUser ? "justify-end" : "justify-start"
-            } w-full`}
-          >
-            {!isCurrentUser &&
-              conversationType === "direct" &&
-              renderAvatar(senderInfo, true, senderId)}
-
-            <div className="flex flex-col max-w-[70%]">
-              <div
-                className={`
-                  rounded-lg p-3
-                  ${
-                    isCurrentUser
-                      ? "bg-green-100 text-base-content rounded-br-none ml-auto"
-                      : "bg-base-200 rounded-bl-none"
-                  }
-                `}
-              >
-                <p>{renderHighlightedSearchText(parsedMessage.personalMessage, searchQuery)}</p>
-                <div
-                  className={`
-                    flex justify-end items-center text-xs mt-1
-                    ${
-                      isCurrentUser
-                        ? "text-base-content/60"
-                        : "text-base-content/50"
-                    }
-                  `}
-                >
-                  <span>{formatLocalTime(message.createdAt)}</span>
-                  {
-                    <ReadReceipt
-                      message={message}
-                      isCurrentUser={isCurrentUser}
-                      conversationType={conversationType}
-                      teamMembers={teamMembers}
-                      currentUserId={currentUserId}
-                      getReadByTooltip={getReadByTooltip}
-                    />
-                  }
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderApplicationCancelledMessage - Neutral grey theme
-  // =============================================================================
-  const renderApplicationCancelledMessage = (
-    message,
-    parsedMessage,
-    isCurrentUser,
-  ) => {
-    const messageText = isCurrentUser ? (
-      <>
-        You cancelled your application for{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />
-        . Want to tell them why in this chat?
-      </>
-    ) : (
-      <>
-        <MentionById
-          userId={parsedMessage.applicantId}
-          name={parsedMessage.applicantName}
-        />{" "}
-        cancelled their application for{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />
-        . Want to reach out to them in this chat?
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <CircleX size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderRoleChangedMessage - Dynamic theme based on new role
-  // =============================================================================
-  const renderRoleChangedMessage = (message, parsedMessage) => {
-    const isPromotion = parsedMessage.newRole === "admin";
-    const newRole = parsedMessage.newRole;
-
-    const getRoleBannerClass = (role) => {
-      switch (role) {
-        case "owner":
-          return "event-banner--owner";
-        case "admin":
-          return "event-banner--admin";
-        case "member":
-        default:
-          return "event-banner--member";
-      }
-    };
-
-    const getRoleIcon = (role) => {
-      switch (role) {
-        case "owner":
-          return Crown;
-        case "admin":
-          return Shield;
-        case "member":
-        default:
-          return User;
-      }
-    };
-
-    const bannerClass = getRoleBannerClass(newRole);
-    const RoleIcon = getRoleIcon(newRole);
-    const isChangerCurrentUser = isCurrentViewer(parsedMessage.changerId);
-    const isMemberCurrentUser = isCurrentViewer(parsedMessage.memberId);
-
-    const messageText = isChangerCurrentUser ? (
-      isPromotion ? (
-        <>
-          You promoted{" "}
-          <MentionById
-            userId={parsedMessage.memberId}
-            name={parsedMessage.memberName}
-          />{" "}
-          to Admin in{" "}
-          <TeamMentionById
-            teamId={parsedMessage.teamId}
-            name={parsedMessage.teamName}
-          />
-          .
-        </>
-      ) : (
-        <>
-          You changed{" "}
-          <MentionById
-            userId={parsedMessage.memberId}
-            name={parsedMessage.memberName}
-          />
-          {"'s"} role to Member in{" "}
-          <TeamMentionById
-            teamId={parsedMessage.teamId}
-            name={parsedMessage.teamName}
-          />
-          .
-        </>
-      )
-    ) : isMemberCurrentUser ? (
-      isPromotion ? (
-      <>
-        You were promoted to Admin in{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        by{" "}
-        <MentionById
-          userId={parsedMessage.changerId}
-          name={parsedMessage.changerName}
-        />
-        .
-      </>
-      ) : (
-        <>
-          Your role in{" "}
-          <TeamMentionById
-            teamId={parsedMessage.teamId}
-            name={parsedMessage.teamName}
-          />{" "}
-          was changed to Member by{" "}
-          {userMentionOrYou(parsedMessage.changerId, parsedMessage.changerName)}
-          .
-        </>
-      )
-    ) : (
-      <>
-        {possessiveUserMentionOrYour(
-          parsedMessage.memberId,
-          parsedMessage.memberName,
-        )}{" "}
-        role in{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        was {isPromotion ? "changed to Admin" : "changed to Member"} by{" "}
-        {userMentionOrYou(parsedMessage.changerId, parsedMessage.changerName)}
-        .
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className={`event-banner ${bannerClass} mb-3`}>
-          <span className="text-sm font-medium event-message-text">
-            <RoleIcon size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-            {isPromotion && isMemberCurrentUser && (
-              <PartyPopper size={16} className="event-inline-icon ml-1" />
-            )}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderOwnershipTeamMessage - Pink owner theme (team chat)
-  // =============================================================================
-  const renderOwnershipTeamMessage = (message, parsedMessage) => {
-    const isPreviousOwnerCurrentUser = isCurrentViewer(parsedMessage.prevOwnerId);
-    const isNewOwnerCurrentUser = isCurrentViewer(parsedMessage.newOwnerId);
-    const messageText = isPreviousOwnerCurrentUser ? (
-      <>
-        You transferred ownership to{" "}
-        {userMentionOrYou(parsedMessage.newOwnerId, parsedMessage.newOwnerName)}
-      </>
-    ) : isNewOwnerCurrentUser ? (
-      <>
-        {userMentionOrYou(parsedMessage.prevOwnerId, parsedMessage.prevOwnerName, {
-          capitalized: true,
-        })}{" "}
-        transferred ownership to you
-      </>
-    ) : (
-      <>
-        {userMentionOrYou(parsedMessage.prevOwnerId, parsedMessage.prevOwnerName, {
-          capitalized: true,
-        })}{" "}
-        transferred ownership to{" "}
-        {userMentionOrYou(parsedMessage.newOwnerId, parsedMessage.newOwnerName)}
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--owner mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <Crown size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  // =============================================================================
-  // renderOwnershipTransferredMessage - Pink owner theme (DM)
-  // =============================================================================
-  const renderOwnershipTransferredMessage = (
-    message,
-    parsedMessage,
-    isCurrentUser,
-  ) => {
-    const messageText = isCurrentUser ? (
-      <>
-        You transferred team ownership of{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        to <Mention name={parsedMessage.newOwnerName} />.
-      </>
-    ) : (
-      <>
-        <MentionById
-          userId={parsedMessage.prevOwnerId}
-          name={parsedMessage.prevOwnerName}
-        />{" "}
-        transferred ownership of{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        to you. Congratulations!
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--owner mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <Crown size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-            <PartyPopper size={16} className="event-inline-icon ml-1" />
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
-
-  /**
-   * Render a member removed message with special formatting
-   */
-  const renderMemberRemovedMessage = (
-    message,
-    parsedMessage,
-    isCurrentUser,
-  ) => {
-    const messageText = isCurrentUser ? (
-      <>
-        You removed{" "}
-        <MentionById
-          userId={parsedMessage.memberId}
-          name={parsedMessage.memberName}
-        />{" "}
-        from{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />
-        .
-      </>
-    ) : (
-      <>
-        You were removed from{" "}
-        <TeamMentionById
-          teamId={parsedMessage.teamId}
-          name={parsedMessage.teamName}
-        />{" "}
-        by{" "}
-        <MentionById
-          userId={parsedMessage.removerId}
-          name={parsedMessage.removerName}
-        />
-        . Want to reach out to them in this chat?
-      </>
-    );
-
-    return (
-      <div className="flex flex-col items-center w-full my-4">
-        <div className="event-banner event-banner--neutral mb-3">
-          <span className="text-sm font-medium event-message-text">
-            <UserMinus size={16} className="event-inline-icon mr-1" />
-            {highlightEventContent(messageText)}
-          </span>
-        </div>
-
-        <div className="text-xs text-base-content/50">
-          {formatLocalTime(message.createdAt)}
-        </div>
-      </div>
-    );
-  };
 
   // --------------------------------------------
   // NO MESSAGES STATE
